damn, now it really runs

+++ project-rust/src/modeler/components/process.rs
@@ -37,6 +37,6 @@ pub fn get_summary(e: &Element, nearest_tnext: String) -> String {
 }
 
 pub fn do_statistics(e: &mut Element, delta: f64) {
-    e.state_sum += e.state as u32 * delta as u32;
+    e.state_sum += (e.state as f64 * delta) as u32;
     e.mean_queue += e.queue as f64 * delta;
 }
